Move types into @types/index.ts
